rename events namespace (bc break)
