+++ src/Events/CommandBusAwareEventDispatcher.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace Damejidlo\EventBus;
+namespace Damejidlo\Events;
 
 use Damejidlo\MessageBus\Middleware\IsCurrentlyHandlingAwareMiddleware;
