+++ tests/Integration/Fixtures/CreateInvoiceOnOrderPlaced.php
@@ -2,7 +2,7 @@
 
 namespace DamejidloTests\Integration\Fixtures;
 
-use Damejidlo\EventBus\IEventSubscriber;
+use Damejidlo\Events\IEventSubscriber;
