+++ tests/Integration/SynchronousEventsDispatchingTest.phpt
@@ -12,10 +12,10 @@ require_once __DIR__ . '/../bootstrap.php';
 use Damejidlo\CommandBus\CommandBus;
 use Damejidlo\CommandBus\ICommandBus;
 use Damejidlo\CommandBus\NewEntityId;
-use Damejidlo\EventBus\CommandBusAwareEventDispatcher;
-use Damejidlo\EventBus\EventBus;
-use Damejidlo\EventBus\InMemoryEventQueue;
-use Damejidlo\EventBus\SynchronousEventDispatcher;
+use Damejidlo\Events\CommandBusAwareEventDispatcher;
+use Damejidlo\Events\EventBus;
+use Damejidlo\Events\InMemoryEventQueue;
+use Damejidlo\Events\SynchronousEventDispatcher;
 use Damejidlo\MessageBus\Handling\Implementation\ArrayMapHandlerProvider;
 use Damejidlo\MessageBus\Handling\Implementation\ArrayMapHandlerTypesResolver;
 use Damejidlo\MessageBus\Handling\Implementation\HandlerInvoker;
